requirements.txt
flask==2.3.1
gunicorn
groq
joblib
scikit-learn
python-telegram-bot==20.*   # or latest v20.x